# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ lib/ThreadsManager.h
@@ -146,7 +146,7 @@ public:
 #ifdef _WIN32
         SYSTEM_INFO si;
         GetSystemInfo(&si);
-        count = si.dwNumberOfProcessors;
+        count = (int)si.dwNumberOfProcessors;
 #else
         count = static_cast<int>(sysconf(_SC_NPROCESSORS_CONF));
 #endif
